Move ViewModelLocator to ViewModels folder
Issue #158
May not be a simple merge with #165 due to files having moved
