+++ templates/_composition/MVVMLight/Page/ViewModels/ViewModelLocator_postaction.cs
@@ -1,0 +1,20 @@
+﻿namespace RootNamespace.ViewModels
+{
+    public class ViewModelLocator
+    {
+        public ViewModelLocator()
+        {
+            //^^
+            Registeruct.ItemName(navigationService);
+        }
+
+        public uct.ItemNameViewModel uct.ItemNameViewModel => ServiceLocator.Current.GetInstance<uct.ItemNameViewModel>();
+        //{[{
+        public void Registeruct.ItemName(NavigationServiceEx navigationService)
+        {
+            SimpleIoc.Default.Register<uct.ItemNameViewModel>();
+            navigationService.Configure(typeof(uct.ItemNameViewModel).FullName, typeof(uct.ItemNamePage));
+        }
+        //}]}
+    }
+}
