--- templates/_composition/MVVMLight/Project.SplitView/ViewModelLocator_postaction.cs
@@ -1,11 +1,0 @@
-﻿public class ViewModelLocator
-{
-    public ViewModelLocator()
-    {
-        //^^
-        RegisterShell();
-    }
-
-    public ShellViewModel ShellViewModel => ServiceLocator.Current.GetInstance<ShellViewModel>();
-    public void RegisterShell() => SimpleIoc.Default.Register<ShellViewModel>();
-}
