+++ templates/_composition/MVVMLight/Project.SplitView/ViewModels/ViewModelLocator_postaction.cs
@@ -1,0 +1,14 @@
+﻿namespace RootNamespace.ViewModels
+{
+    public class ViewModelLocator
+    {
+        public ViewModelLocator()
+        {
+            //^^
+            RegisterShell();
+        }
+
+        public ShellViewModel ShellViewModel => ServiceLocator.Current.GetInstance<ShellViewModel>();
+        public void RegisterShell() => SimpleIoc.Default.Register<ShellViewModel>();
+    }
+}
