--- templates/_composition/MVVMLight/Project.TabbedPivot/ViewModelLocator_postaction.cs
@@ -1,17 +1,0 @@
-﻿public class ViewModelLocator
-{
-    public ViewModelLocator()
-    {
-        //^^
-        RegisterPivotView(navigationService);
-    }
-
-    public PivotViewModel PivotViewModel => ServiceLocator.Current.GetInstance<PivotViewModel>();
-    //{[{
-    public void RegisterPivotView(NavigationServiceEx navigationService)
-    {
-        SimpleIoc.Default.Register<PivotViewModel>();
-        navigationService.Configure(typeof(PivotViewModel).FullName, typeof(PivotPage));
-    }
-    //}]}
-}
