+++ templates/_composition/MVVMLight/Project.TabbedPivot/ViewModels/ViewModelLocator_postaction.cs
@@ -1,0 +1,20 @@
+﻿namespace RootNamespace.ViewModels
+{
+    public class ViewModelLocator
+    {
+        public ViewModelLocator()
+        {
+            //^^
+            RegisterPivotView(navigationService);
+        }
+
+        public PivotViewModel PivotViewModel => ServiceLocator.Current.GetInstance<PivotViewModel>();
+        //{[{
+        public void RegisterPivotView(NavigationServiceEx navigationService)
+        {
+            SimpleIoc.Default.Register<PivotViewModel>();
+            navigationService.Configure(typeof(PivotViewModel).FullName, typeof(PivotPage));
+        }
+        //}]}
+    }
+}
