+++ templates/_composition/MVVMLight/Project/App_postaction.xaml
@@ -1,10 +1,10 @@
 <Application
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:RootNamespace"
+    xmlns:vms="using:RootNamespace.ViewModels"
     RequestedTheme="Light">
     <Application.Resources>
         <ResourceDictionary>
-            <local:ViewModelLocator x:Key="Locator" />
+            <vms:ViewModelLocator x:Key="Locator" />
         </ResourceDictionary>
     </Application.Resources>
 </Application>
